+++ Source/Extensions/Blazorise.DataGrid/Contexts/FilterContext.cs
@@ -1,0 +1,45 @@
+﻿namespace Blazorise.DataGrid
+{
+    /// <summary>
+    /// Context for editors in datagrid filter section.
+    /// </summary>
+    public class FilterContext
+    {
+        #region Members
+
+        private event FilterChangedEventHandler SearchValueChanged;
+
+        public delegate void FilterChangedEventHandler( object value );
+
+        #endregion
+
+        #region Methods
+
+        public void Subscribe( FilterChangedEventHandler listener )
+        {
+            SearchValueChanged += listener;
+        }
+
+        public void Unsubscribe( FilterChangedEventHandler listener )
+        {
+            SearchValueChanged -= listener;
+        }
+
+        public void TriggerFilterChange( object value )
+        {
+            SearchValue = value;
+            SearchValueChanged?.Invoke( value );
+        }
+
+        #endregion
+
+        #region Properties
+
+        /// <summary>
+        /// Gets or sets the filter value.
+        /// </summary>
+        public object SearchValue { get; set; }
+
+        #endregion
+    }
+}
